+++ udata_schema_gouvfr/tasks.py
@@ -6,15 +6,13 @@ import logging
 
 import requests
 
+from flask import current_app
 from udata.tasks import job
 from udata.core.dataset.models import get_resource, ResourceSchema
 
 log = logging.getLogger(__name__)
 
 
-IRVE_STABLE_DATASET_URL = (
-    "https://www.data.gouv.fr/fr/datasets/r/50625621-18bd-43cb-8fde-6b8c24bdabb3"
-)
 IRVE_SCHEMA = "etalab/schema-irve"
 
 
@@ -22,7 +20,7 @@ def find_sources():
     """
     Find all (dataset, resource) listed in the consolidated IRVE resource
     """
-    response = requests.get(IRVE_STABLE_DATASET_URL)
+    response = requests.get(current_app.config.get('SCHEMA_GOUVFR_IRVE_STABLE_RESOURCE_URL'))
     response.raise_for_status()
 
     sources = set()
@@ -59,7 +57,7 @@ def set_irve_schemas(sources):
         dataset_slug, resource_id = re.match(pattern, source).groups()
         resource = get_resource(uuid.UUID(resource_id))
         if not resource:
-            raise ValueError(f'Cannot find a resource with ID {resource_id}')
+            log.warning(f'Cannot find a resource with ID {resource_id}. It may have been deleted.')
         resource.schema = IRVE_SCHEMA
         resource.save()
         log.info(f'Set the {IRVE_SCHEMA} on resource {resource_id} from dataset {dataset_slug}')