+++ udata_schema_gouvfr/__init__.py
@@ -4,5 +4,5 @@ udata-schema-gouvfr
 Integration with schema.data.gouv.fr
 '''
 
-__version__ = '1.0.1.dev'
+__version__ = '1.1.0'
 __description__ = 'Integration with schema.data.gouv.fr'