+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Add task to set the schema attribute to the IRVE schema on multiple resources [#3](https://github.com/etalab/udata-schema-gouvfr/pull/3)
+- Add a job to set the `schema` attribute to the IRVE schema on multiple resources [#3](https://github.com/etalab/udata-schema-gouvfr/pull/3)
 
 ## 1.0.0 (2020-08-05)