+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 1.1.0 (2020-08-25)
 
 - Add a job to set the `schema` attribute to the IRVE schema on multiple resources [#3](https://github.com/etalab/udata-schema-gouvfr/pull/3)